model/requirements.txt
# Core ML
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.11.0
bitsandbytes>=0.43.3

# Data
datasets>=2.20.0
pillow>=10.3.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
playwright>=1.44.0

# Evaluation
sacrebleu>=2.4.0

# Misc
sentencepiece>=0.2.0
qwen-vl-utils>=0.0.8
